+++ tests/mod.rs
@@ -57,6 +57,19 @@ fn calibrate_offset_slow_ok() {
 }
 
 #[test]
+fn clear_alarm_interrupt_ok() {
+    let expectations = [
+        I2cTransaction::read(PCF8523_I2C_ADDRESS, [0b0].to_vec()),
+        i2c_reg_read(PCF8523_CONTROL_2, 0b0010_1011),
+        i2c_reg_write(PCF8523_CONTROL_2, 0b0010_0011),
+    ];
+    let mut i2c = I2cMock::new(&expectations);
+    let mut driver = Pcf8523::new(&mut i2c).unwrap();
+    driver.clear_alarm_interrupt().unwrap();
+    i2c.done();
+}
+
+#[test]
 fn clear_second_interrupt_ok() {
     let expectations = [
         I2cTransaction::read(PCF8523_I2C_ADDRESS, [0b0].to_vec()),