+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable/disable correction interrupt
 - Enable/disable/clear second interrupt
+- Clear alarm interrupt
 
 ## [0.2.0] - 2025-08-23